phone-app: Milestone 2 ONNX backends + Android APK builds
Wires real-model inference paths for the v2 phone app while keeping the
stub detector as the fallback. Confirms the Android debug APK assembles
end-to-end on the Linux VM (~82 MB; CameraX + ONNX Runtime native libs
included). Confirms `:composeApp:desktopTest` is green for the shared
core unit tests.

- OnnxRuntimeAndroidBackend: tries NNAPI EP with FP16 first, falls back
  to CPU. Records the actual delegate that succeeded so the HUD doesn't
  lie about acceleration.
- OnnxRuntimeJvmBackend: CPU-only desktop equivalent for the still-image
  harness.
- AndroidModelLoader: resolves a ModelSpec to an absolute file path —
  external files dir → cache → APK assets (copied once) — so the model
  can be sideloaded via adb push or bundled into the APK without
  committing weights anywhere in the repo.
- MainActivity.chooseBackend: picks ONNX when the YOLOX-nano poop model
  is found, falls back to StubDetectorBackend otherwise.
- Desktop Main: --model=<path> / --model-id=<id> args + stub fallback.
- build.gradle.kts: -Xexpect-actual-classes to silence the KMP beta
  warning on BuildInfo.
- DetectorBackend.kt: fixed a Float/Double mismatch in the stub jitter
  math that broke the first compile.

Validation that ran inside the VM:
- ./gradlew :composeApp:desktopTest      → SUCCESS, all unit tests pass
- ./gradlew :composeApp:assembleDebug    → SUCCESS, APK at
  composeApp/build/outputs/apk/debug/composeApp-debug.apk
- ./gradlew :composeApp:compileKotlinDesktop :composeApp:compileDebugKotlinAndroid → SUCCESS

What still needs a phone:
- adb install on Pixel 5
- live FPS / inference-latency measurement against GOAL.md targets
- pulling failure-case captures back

Full operator checklist in tpl/shitspotter-phone-app/docs/001_build_run_validate.md
and run journal in dev/journals/2026-05-10_phone_app_kmp_scaffold.md.

+++ tpl/shitspotter-phone-app/composeApp/src/commonMain/kotlin/io/kitware/shitspotter/core/DetectorBackend.kt
@@ -46,8 +46,8 @@ class StubDetectorBackend(
         val n = (counter++ % 600L)
         val w = frame.width.toFloat()
         val h = frame.height.toFloat()
-        val cx = w * (0.4f + 0.2f * sineApprox(n / 60.0))
-        val cy = h * (0.5f + 0.15f * cosineApprox(n / 45.0))
+        val cx = w * (0.4f + 0.2f * sineApprox(n / 60.0).toFloat())
+        val cy = h * (0.5f + 0.15f * cosineApprox(n / 45.0).toFloat())
         val bw = w * 0.18f
         val bh = h * 0.22f
         val fakeBox = BoundingBox(
